fix long running conversation id acceptance test
