+++ src/NServiceBus.AcceptanceTests/Core/Pipeline/When_setting_conversation_id.cs
@@ -3,8 +3,6 @@
     using System;
     using System.Threading.Tasks;
     using AcceptanceTesting;
-    using AcceptanceTesting.Customization;
-    using AcceptanceTesting.Support;
     using EndpointTemplates;
     using NUnit.Framework;
 
@@ -28,27 +26,6 @@
                 .Run();
 
             Assert.AreEqual(customConversationId, context.ReceivedConversationId);
-        }
-
-        [Test]
-        public void Should_throw_when_incoming_conversation_id_available()
-        {
-            var initialConversationId = Guid.NewGuid().ToString();
-
-            var exception = Assert.ThrowsAsync<MessageFailedException>(() => Scenario.Define<Context>()
-                .WithEndpoint<IntermediateEndpoint>(e => e
-                    .When(s =>
-                    {
-                        var options = new SendOptions();
-                        options.RouteToThisEndpoint();
-                        options.SetHeader(Headers.ConversationId, initialConversationId);
-                        return s.Send(new IntermediateMessage(), options);
-                    }))
-                .WithEndpoint<ReceivingEndpoint>()
-                .Done(c => !string.IsNullOrEmpty(c.ReceivedConversationId))
-                .Run());
-
-            StringAssert.Contains($"Cannot set the {Headers.ConversationId} header to 'intermediate message header' as it cannot override the incoming header value ('{initialConversationId}').", exception.InnerException.Message);
         }
 
         class Context : ScenarioContext
@@ -80,30 +57,7 @@
             }
         }
 
-        class IntermediateEndpoint : EndpointConfigurationBuilder
-        {
-            public IntermediateEndpoint()
-            {
-                EndpointSetup<DefaultServer>(e => e.ConfigureTransport().Routing()
-                    .RouteToEndpoint(typeof(MessageWithConversationId), typeof(ReceivingEndpoint)));
-            }
-
-            public class IntermediateMessageHandler : IHandleMessages<IntermediateMessage>
-            {
-                public Task Handle(IntermediateMessage message, IMessageHandlerContext context)
-                {
-                    var options = new SendOptions();
-                    options.SetHeader(Headers.ConversationId, "intermediate message header");
-                    return context.Send(new MessageWithConversationId(), options);
-                }
-            }
-        }
-
         public class MessageWithConversationId : ICommand
-        {
-        }
-
-        public class IntermediateMessage : ICommand
         {
         }
     }
